+++ src/reskyblock/models/bazaar.py
@@ -37,4 +37,4 @@ class Bazaar(msgspec.Struct, rename="camel"):
     success: bool
     last_updated: int
     products: dict[str, Product]
-    received_at: int = 0
+    received_at: float = 0
